# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: Naiad API
  version: "1.0.0"
  description: Garden irrigation controller — REST API
servers:
  - url: /api
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    ScheduleSummary:
      type: object
      required: [days, times, cron]
      properties:
        days:   { type: array, items: { type: integer }, description: "ISO weekdays 1=Mon..7=Sun; empty = every day" }
        times:  { type: array, items: { type: string }, description: "Clock times HH:MM" }
        cron:   { type: string, nullable: true, description: "Advanced cron override; set only when active" }
...
    NotificationCategory:
      type: string
      enum: [start, skip, abort, reminder]
    NotifyTarget:
      type: object
      required: [service, categories, quiet, platform]
      properties:
        service: { type: string }
        categories:
          type: array
          items: { $ref: '#/components/schemas/NotificationCategory' }
        quiet: { type: boolean }
        platform:
          type: string
          enum: [auto, ios, android]
    HAConfigPublic:
      type: object
      required: [url, notify_targets]
      properties:
        url: { type: string }
        notify_targets:
          type: array
          items: { $ref: '#/components/schemas/NotifyTarget' }
    MQTTConfig:
      type: object
      required: [enabled, host, port, username, client_id, discovery_prefix, base_topic, password_set]
      properties:
        enabled: { type: boolean }
        host: { type: string }
        port: { type: integer }
        username: { type: string }
        client_id: { type: string }
        discovery_prefix: { type: string }
        base_topic: { type: string }
        password_set:
          type: boolean
          description: Whether an environment-managed MQTT password is configured. The password value is never exposed.
    MQTTConfigInput:
      type: object
      required: [enabled, host, port, username, client_id, discovery_prefix, base_topic]
      properties:
        enabled: { type: boolean }
        host: { type: string }
        port: { type: integer }
        username: { type: string }
        client_id: { type: string }
        discovery_prefix: { type: string }
        base_topic: { type: string }
        password_set:
          type: boolean
          description: Accepted for GET→edit→PUT round-trips and ignored by the backend.
    AutoLoginTrigger:
      type: object
      required: [url_param, trusted_referers, trusted_ips]
      properties:
        url_param: { type: string }
        trusted_referers:
          type: array
          items: { type: string }
        trusted_ips:
          type: array
          items: { type: string }
    AutoLoginConfig:
      type: object
      required: [enabled, trigger]
      properties:
        enabled: { type: boolean }
        trigger: { $ref: '#/components/schemas/AutoLoginTrigger' }
    ForwardHeaderConfig:
      type: object
      required: [header, trusted_proxies]
      properties:
        header: { type: string }
        trusted_proxies:
          type: array
          items: { type: string }
    AuthConfig:
      type: object
      required: [mode, forward_header, auto_login, frame_ancestors, password_set]
      properties:
        mode:
          type: string
          enum: [password, forward_header, none]
        forward_header: { $ref: '#/components/schemas/ForwardHeaderConfig' }
        auto_login: { $ref: '#/components/schemas/AutoLoginConfig' }
        frame_ancestors:
          type: array
          items: { type: string }
        password_set:
          type: boolean
          description: Whether an environment-managed password is configured. The password value is never exposed.
    AuthConfigInput:
      type: object
      required: [mode, forward_header, auto_login, frame_ancestors]
      properties:
        mode:
          type: string
          enum: [password, forward_header, none]
        forward_header: { $ref: '#/components/schemas/ForwardHeaderConfig' }
        auto_login: { $ref: '#/components/schemas/AutoLoginConfig' }
        frame_ancestors:
          type: array
          items: { type: string }
        password_set:
          type: boolean
          description: Accepted for GET→edit→PUT round-trips and ignored by the backend.
    SensorsConfig:
      type: object
      required: [rain, wind, season, temperature, temperature_max, precipitation_prob_today, precipitation_prob_tomorrow, precipitation_today, precipitation_tomorrow, precipitation_actual]
      properties:
        rain: { type: string }
        wind: { type: string }
        season: { type: string }
        temperature: { type: string }
        temperature_max: { type: string }
        precipitation_prob_today: { type: string }
        precipitation_prob_tomorrow: { type: string }
        precipitation_today: { type: string }
        precipitation_tomorrow: { type: string }
        precipitation_actual: { type: string }
    ZoneConfig:
      type: object
      required: [label, switch, flow_lph, staircase_enabled, staircase_min]
      properties:
        label: { type: string }
        switch: { type: string }
        flow_lph: { type: number, exclusiveMinimum: 0 }
        staircase_enabled: { type: boolean }
        staircase_min: { type: number }
    SequenceColorKey:
      type: string
      enum: [green, sand, purple, slate, blue, rose]
    SequenceConfig:
      type: object
      required: [label, zones, basis_min_per_zone, range, watchdog_min, schedule, enabled, wind_blocks, color]
      properties:
        label: { type: string }
        zones:
          type: array
          items: { type: string }
        basis_min_per_zone: { type: number, exclusiveMinimum: 0 }
        range:
          type: array
          minItems: 2
          maxItems: 2
          prefixItems:
            - { type: number }
            - { type: number }
        watchdog_min: { type: integer, exclusiveMinimum: 0 }
        schedule: { $ref: '#/components/schemas/ScheduleSummary' }
        enabled: { type: boolean }
        wind_blocks: { type: boolean }
        color:
          nullable: true
          oneOf:
            - { $ref: '#/components/schemas/SequenceColorKey' }
            - { type: 'null' }
    TempFactorConfig:
      type: object
      required: [formula, basis_c, pct_per_c, min_pct, max_pct]
      properties:
        formula:
          type: string
          enum: [linear]
        basis_c: { type: number }
        pct_per_c: { type: number }
        min_pct: { type: integer, minimum: 0 }
        max_pct: { type: integer, minimum: 0 }
    RainFactorConfig:
      type: object
      required: [mode, forecast_days, threshold_prob, reduce_above_mm, zero_above_mm, forecast_decay, water_balance_days, water_balance_decay, peak_tomorrow, confirm_with_rain_sensor]
      properties:
        mode:
          type: string
          enum: [forecast, water_balance]
        forecast_days: { type: integer, minimum: 1 }
        threshold_prob: { type: integer, minimum: 0, maximum: 100 }
        reduce_above_mm: { type: number }
        zero_above_mm: { type: number }
        forecast_decay: { type: number, minimum: 0, maximum: 1 }
        water_balance_days: { type: integer, minimum: 1, maximum: 14 }
        water_balance_decay: { type: number, minimum: 0, maximum: 1 }
        peak_tomorrow: { type: boolean }
        confirm_with_rain_sensor: { type: boolean }
    FactorsConfig:
      type: object
      required: [temp, rain]
      properties:
        temp: { $ref: '#/components/schemas/TempFactorConfig' }
        rain: { $ref: '#/components/schemas/RainFactorConfig' }
    NotificationsConfig:
      type: object
      required: [evening_reminder_cron, queue_max_hours]
      properties:
        evening_reminder_cron: { type: string }
        queue_max_hours: { type: number, minimum: 0 }
    ConfigDoc:
      type: object
      required: [ha, mqtt, auth, sensors, zones, sequences, factors, notifications, timezone, sequence_colors_enabled, restart_required]
      properties:
        ha: { $ref: '#/components/schemas/HAConfigPublic' }
        mqtt: { $ref: '#/components/schemas/MQTTConfig' }
        auth: { $ref: '#/components/schemas/AuthConfig' }
        sensors: { $ref: '#/components/schemas/SensorsConfig' }
        zones:
          type: object
          additionalProperties: { $ref: '#/components/schemas/ZoneConfig' }
        sequences:
          type: object
          additionalProperties: { $ref: '#/components/schemas/SequenceConfig' }
        factors: { $ref: '#/components/schemas/FactorsConfig' }
        notifications: { $ref: '#/components/schemas/NotificationsConfig' }
        timezone: { type: string }
        sequence_colors_enabled: { type: boolean }
        restart_required: { type: boolean }
    ConfigUpdateRequest:
      type: object
      required: [ha, mqtt, auth, sensors, zones, sequences, factors, notifications, timezone, sequence_colors_enabled]
      properties:
        ha: { $ref: '#/components/schemas/HAConfigPublic' }
        mqtt: { $ref: '#/components/schemas/MQTTConfigInput' }
        auth: { $ref: '#/components/schemas/AuthConfigInput' }
        sensors: { $ref: '#/components/schemas/SensorsConfig' }
        zones:
          type: object
          additionalProperties: { $ref: '#/components/schemas/ZoneConfig' }
        sequences:
          type: object
          additionalProperties: { $ref: '#/components/schemas/SequenceConfig' }
        factors: { $ref: '#/components/schemas/FactorsConfig' }
        notifications: { $ref: '#/components/schemas/NotificationsConfig' }
        timezone: { type: string }
        sequence_colors_enabled: { type: boolean }
        restart_required:
          type: boolean
          description: Accepted for GET→edit→PUT round-trips and ignored by the backend.
    EntityInfo:
      type: object
      required: [entity_id, friendly_name, state, domain]
      properties:
        entity_id: { type: string }
        friendly_name: { type: string, nullable: true }
        state: { type: string }
        domain: { type: string }
    EntitiesResponse:
      type: object
      required: [entities]
      properties:
        entities:
          type: array
          items: { $ref: '#/components/schemas/EntityInfo' }
    ServicesResponse:
      type: object
      required: [services]
      properties:
        services:
          type: array
          items: { type: string }
    TestNotifyResponse:
      type: object
      required: [sent, targets]
      properties:
        sent: { type: integer }
        targets:
          type: array
          items: { type: string }

    # ── Auth ──────────────────────────────────────────────────────────────────
...
paths:
...
  /config:
    get:
      operationId: getConfiguration
      summary: Public configuration document with secrets redacted
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ConfigDoc' }
    put:
      operationId: replaceConfiguration
      summary: Replace the configuration document
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ConfigUpdateRequest' }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ConfigDoc' }
        '409':
          description: Cannot change configuration while valve operations are active
        '422':
          description: Invalid configuration
...
  /config/import:
    post:
      operationId: importConfiguration
      summary: Import a YAML or JSON configuration document
      requestBody:
        required: true
        content:
          application/x-yaml:
            schema: { type: string }
          application/json:
            schema: { type: object }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ConfigDoc' }
        '400':
          description: Uploaded config could not be parsed
        '409':
          description: Cannot change configuration while valve operations are active
        '422':
          description: Invalid configuration
  /config/entities:
    get:
      operationId: listEntities
      summary: List cached Home Assistant entities for UI pickers
      parameters:
        - name: domain
          in: query
          schema: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EntitiesResponse' }
  /config/services:
    get:
      operationId: listServices
      summary: List Home Assistant services for UI pickers
      parameters:
        - name: domain
          in: query
          schema: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ServicesResponse' }
  /config/test-notify:
    post:
      operationId: testNotify
      summary: Send a test notification to configured notify targets
      parameters:
        - name: service
          in: query
          schema: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TestNotifyResponse' }
        '400':
          description: No notify targets configured
        '404':
          description: Notify target not found
        '502':
          description: Notification delivery failed